Fixes #676
